requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# DeepSeek / AI
openai==1.51.2
langchain==0.3.7
langchain-community==0.3.5

# Telegram
aiogram==3.13.1

# WhatsApp
httpx==0.27.2

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3

# Redis
redis==5.1.1

# Vector DB
qdrant-client==1.11.3

# Document parsing
pypdf==5.0.1
python-docx==1.1.2
markdown==3.7

# Deployment
docker==7.1.0
